requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML
openai==1.10.0
groq==0.4.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-groq==0.0.1
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.7.4
chromadb==0.4.22
sentence-transformers==3.3.1

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
